[mini_cluster] introduce 'builtin' clock source
This patch introduces the built-in NTP client into the external
minicluster test scaffolding.  With this patch, it's possible
to synchronize the built-in NTP client of the minicluster's masters
and tablet servers with chronyd NTP server which is run as a part
of the minicluster as well.  The latter runs in server-only mode
(i.e. not driving the system clock), using the machine's clock
as a reference (for details, see 'man chronyd', '-x' option).

Change-Id: I5c334ae6fa1fb12b033de7f8e8584b8dd3aa2d32
Reviewed-on: http://gerrit.cloudera.org:8080/14227
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>