KUDU-2069 p4: stop replication from failed servers in maintenance mode
When determining whether a replica needs to be added, we may now
consider a set of of UUIDs that are allowed to be in a "bad" state while
not counting towards being under-replicated.

Since the goal of maintenance mode is to cope with unexpected failures,
"healthy" movement, e.g. through tooling that uses REPLACE and PROMOTE
tagging, is still allowed to and from tservers in maintenance mode.

Testing:
- a unit test is added to exercise the new quorum logic to ignore
  certain failed UUIDs, taking into account REPLACE and PROMOTE replicas
- integration tests are added to test:
  - behavior with RF=3 through restarts of the master
  - behavior when running move_replica tooling
  - behavior with RF=5 with background failures

Change-Id: I9a63b55011d16900c0d27eac0eb75880074204db
Reviewed-on: http://gerrit.cloudera.org:8080/14222
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>