tools: don't open block manager when dumping UUID
The block manager can be expensive to open, even when opening it in
read-only mode. If all a user wants is the server's UUID, all we really
need is to read the instance metadata files.

This patch updates the `kudu fs dump uuid` tool to do just that.

This also moves some words around in the header to try clarifying what
functions are important for startup.

Change-Id: I0acf3b08fc13c169a52d4cfe847efd62bd30cc0a
Reviewed-on: http://gerrit.cloudera.org:8080/14243
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins