deflake HmsCatalogTest.TestMetastoreUuid
In the test, it's possible that the first time we get the UUID, we don't
actually get anything, e.g., because the HMS is still starting up.

If that happens, the first call will respond with a NotSupported error,
even if the HMS supports UUIDs. The later fetch of the UUID may
successfully return a UUID and then hit an assertion failure.

When running the test with a version of Kudu that uses Hive 3.1, the
test failed 7/20 times. With this patch, the test passed 100/100 times.

Change-Id: I34becd77db193554020ca21171482e6ccd7218d6
Reviewed-on: http://gerrit.cloudera.org:8080/14257
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>