KUDU-2946: avoid ServicePool inc/dec in Messenger::QueueInboundCall
Before commit 0ecc2c771, this function took the Messenger lock for what
seemed like longer than necessary. To reduce the size of the critical
section, we had to explicitly take a ServicePool ref, then release it at the
end of the function.

This turned out to have unintended side effects: it's possible for the
release to drop the last ref and destroy the ServicePool. The problem with
this is that QueueInboundCall is called by the reactor thread and destroying
the ServicePool is a blocking operation.

Rather than untangle the morass of _how_ the reactor thread ended up with
the last ServicePool ref, let's just revert this portion of that commit.

Change-Id: Id759617dcb13a2533e9ce071880c43678b700d25
Reviewed-on: http://gerrit.cloudera.org:8080/14259
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Kudu Jenkins