[itest] unify signature of functions using TServerDetails*
This patch updates signatures of test cluster-related
helper functions to use 'const TServerDetails*' when possible.
The motivation for this change was finding that it's possible to use
'const TServerDetails*' argument for some of those, but for others
it would require adding const cast.

There are some functions left that still use non-const TServerDetails*.
Those are eventually call GetLastOpIdForEachReplica(), where one of the
parameters has type of 'const vector<TServerDetails*>&'.

This patch doesn't contain any functional changes.

Change-Id: I7d682be4572f588ee073efdc73d52089d6909dc0
Reviewed-on: http://gerrit.cloudera.org:8080/14269
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>