consensus: remove --raft_attempt_to_replace_replica_without_majority
The flag was introduced for testing, but it seems that testing never
materialized as something automated, except for testing of some quorum
utility functions.

It's an unsafe flag that we don't expect users to ever toggle, so this
patch removes it, along with the testing of the quorum utility
functions.

Change-Id: I36d37c6c3671a524bda229684f4faf60776f2aa5
Reviewed-on: http://gerrit.cloudera.org:8080/14284
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>