KUDU-1938 Add support for VARCHAR pt 1
Introduces the VARCHAR data type to the server. Follow up commits will
add integration to the clients. The VARCHAR type is parameterized with a
length column type attribute similar to DECIMAL's scale and precision.
Internally it's stored as BINARY.

The maximum length for VARCHAR is 65,535 characters. If a value longer
than the "n" is submitted for a VARCHAR(n) column the value is truncated
to "n" characters before persisting the data on the server side.

The maximum length was chosen for compatibility reasons. Apache Impala
has a maximum length of 65,535 *bytes* for VARCHAR and major RDBMSs I
checked also have a similar limits either in characters or bytes, mostly
configurable.

Change-Id: I998982dba93831db91c43a97ce30d3e68c2a4a54
Reviewed-on: http://gerrit.cloudera.org:8080/13760
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>
Reviewed-by: Adar Dembo <adar@cloudera.com>