KUDU-1938 Add non-copy VARCHAR setters pt 3
Apache Impala uses KuduPartialRow API to determine which partition a row
will be inserted to distribute the data between executors optimally.

For this purpose the copy is unnecessary and it should be fast. This
commit adds NoCopyUnsafe variants for this purpose which expect the data
to already be truncated (which it is in Impala's case) and only check
that the value's length is lower than the highest possible upper bound:
val.size() < max_length*4 bytes (the maximum size of an UTF8 character)
to avoid having to count each character manually.

Change-Id: I1f2aba098d649eb94e0314f6606cc33600e8d766
Reviewed-on: http://gerrit.cloudera.org:8080/13928
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins