KUDU-2800 Test long bootstrapping tablet replicas
These tests cover change of consensus state in case of long
bootstrapping of a tablet replica.
The following cases are covered:

1) A tablet replica bootstraps long time. No data modications happen
during that time. The replica is not evicted from the tablet Raft
configuration and joins the quorum after the bootstrapping process is
finished.

2) A tablet replica is shut down for some time. During this time a lot
of data modifications happen, log rolls, the replica is replaced (with
re-replication). The replica restarts and finds out that is has been
evicted from the tablet Raft configuration.

3) A tablet replica bootstraps long time. During this time a lot
of data modifications happen, log rolls, the replica is replaced (with
re-replication). The replica eventually finishes boostrap and finds out
that is has been evicted from the tablet Raft configuration.

Change-Id: Ie1fee42053194f51d7a869ce14788095d6627ed9
Reviewed-on: http://gerrit.cloudera.org:8080/14300
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>