KUDU-2069 p7: add tool to enter/exit maintenance
Adds the following tools:
$ kudu tserver state enter_maintenance <masters> <ts_id>
$ kudu tserver state exit_maintenance <masters> <ts_id>

This opts to have users specify the transition they expect to enact,
rather than the desired end state. This is because I expect the same
tool to be used for decommissioning. When that materializes, it will be
valuable to have users specify their expected starting state to avoid
accidentally overwriting tserver states unintentionally.

It also adds an option into the TSManager to allow the tool to raise an
error when running it with a tserver that doesn't exist. This isn't
always desired behavior of the TSManager because we will reload tserver
states when the master first starts up, at which point tservers may not
be registered yet. It is accessible through a CLI tool optional
argument.

Change-Id: Ifab64dcb5b7b0d171502c9154c83113c82e5972d
Reviewed-on: http://gerrit.cloudera.org:8080/14368
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>