[clock] change in the semantics of TimeService::Init()
This patch introduces a change in the semantics of the TimeService's
initialization method Init().  With this patch, TimeService::Init()
no longer waits for the clock synchronisation to happen for NTP-based
time services.  Users of the TimeService interface should rely on
TimeService::WalltimeWithError() to get the status of the clock
synchronisation after calling TimeService::Init().

This is a follow-up to e110cb88b145edb6d536ae0806e02736f5bf51e5.

Change-Id: Ie7245b2cd2c3001a508235f9e539d7202d3ca994
Reviewed-on: http://gerrit.cloudera.org:8080/14317
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>