test: deflake TestBackgroundFailureDuringMaintenanceMode
We previously weren't accounting for the possibility of a slow
bootstrap, asserting early that we had the expected number of running
replicas.

This also fixes the check for 0 failed replicas. Really, we should have
been checking for explicit healthy/failed states, rather than "not
failed".

Without this, the test failed 6/1000 times in debug mode. With it, it
passed 4997/5000 times (failures due to a different issue).

Change-Id: I67eed008a983cd9ed528496777ffe3d30126a55f
Reviewed-on: http://gerrit.cloudera.org:8080/14341
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>