[mini_cluster] small fix on Mini{Chronyd,Hms}::Start()
This patch makes it possible to start ExternalMiniCluster on a directory
structure originally created by an earlier ExternalMiniCluster, provided
the same set of master RPC endpoints are used in both cases.

This functionality is useful for test scenarios like
MasterMigrationTest.TestEndToEndMigration, when a Kudu master catalog
is preserved between cluster's restarts.

The motivation for this patch was seeing the mentioned test scenario
failing if running with the built-in NTP.  That's because
ExternalMiniCluster does the ephemeral port binding, and then hands off
the port to mini_chronyd.  Since the port binding is ephemeral, next
start usually gets different port bound.  However, upon next start,
mini_chronyd wasn't recreating the configuration file for chronyd
as necessary in MiniChronyd::Start().  I also found that running
the scenario with HMS enabled would fail as well, so I updated
MiniHms::Start() accordingly.

In addition, I updated the ExternalMiniClusterTest.TestBasicOperation
test scenario to cover the updated functionality.

Change-Id: I014eac84386c4d6e4f025afc1082e6a229a97454
Reviewed-on: http://gerrit.cloudera.org:8080/14345
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>