[mini_cluster] set MiniChronydOptions::data_root explicitly
This patch explicitly sets MiniChronydOptions::data_root when adding
test NTP server to the external mini-cluster.  This is necessary to
run the 'test mini_cluster' sub-command of the kudu CLI tool since
the auto-assignment of the 'data_root' field works only when the
mini_cluster is run in the context of a Kudu test.

In addition, this test sets the initial timeout for NTP clock
synchronisation (--ntp_initial_sync_wait_secs) to 10 seconds when
enabling the built-in NTP client.  As it turned out, 2 seconds sometimes
was not enough for TSAN builds.

Change-Id: I1103f7048c2782da7167250962ef19df3488c4ba
Reviewed-on: http://gerrit.cloudera.org:8080/14346
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>