[mini_chronyd] introduce MiniChronyd::address()
Introduced MiniChronyd::address() method to find IP address and port
at which the underlying NTP server is listening for incoming requests.
In case if the NTP server is bound to the wildcard address, use the
loopback address for NTP communication with the server.

Also, updated corresponding call sites to use the new method instead of
accessing MiniChronyd's options.

Prior to this patch, the ToolTest.TestRemoteReplicaCopy test scenario
would fail if running with built-in NTP client enabled.

Change-Id: I62a326a47160aefe9a5806162e6df7c4d6060ba5
Reviewed-on: http://gerrit.cloudera.org:8080/14355
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>