[test] use built-in NTP by default for ExternalMiniCluster
This patch switches all ExternalMiniCluster-based tests to use
the built-in NTP client as the default source for HybridTime timestamps.

A new environment variable is introduced to override the default when
running tests: KUDU_USE_SYSTEM_NTP.  Set it to true if it's desired
to run the ExternalMiniCluster-based scenarios using the machine's local
clock as the source for HybridTime (the clock should be synchronized by
the kernel NTP discipline).

Change-Id: I2a93bc3847ff5b81b40106e5074934e26136320a
Reviewed-on: http://gerrit.cloudera.org:8080/14362
Tested-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Grant Henke <granthenke@apache.org>