tablet_server-test: deflake ScanCorruptedDeltasParamTest
Saw this in a precommit, though I could only repro it 1/10000 times using
dist-test, so it must be pretty rare. Normally scans on a failed tablet will
yield the error already expected in the test, but tablet failure is
asynchronous, so if we're doing snapshot scans, it's possible for the next
scan to see a running tablet, start the scan, and fail within
WaitForSnapshotWithAllCommitted because the async failure process hasn't yet
marked the entire tablet as failed.

  /home/jenkins-slave/workspace/kudu-master/1/src/kudu/tserver/tablet_server-test.cc:1840: Failure
        Expected: resp.error().status().code()
        Which is: 11
  To be equal to: AppStatusPB::ILLEGAL_STATE
        Which is: 9
  Google Test trace:
  /home/jenkins-slave/workspace/kudu-master/1/src/kudu/tserver/tablet_server-test.cc:1839: error {
    code: TABLET_FAILED
    status {
      code: ABORTED
      message: "MVCC is closed"
    }
  }

Change-Id: I51012fa5912e4ce6a2e8f1d370ac387df5bd6db6
Reviewed-on: http://gerrit.cloudera.org:8080/14375
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>