[tests] use system_unsync time source for tests by default
This patch switches all KuduTest-derived tests to use 'system_unsync'
time source by default (note that becomes so for test based on both
internal and external mini-clusters).  The rationale is explained below.

Essentially, there is no need to require NTP-synchronized clock for
a generic test derived from KuduTest.  All the participating Kudu
masters and tablet servers (and their sub-components in case of unit
tests) are run at the same node using the same local wallclock.  Using
the same wallclock makes the synchronisation of clocks unnecessary.

The scenarios that verify the functioning of particular Kudu subsystems
when using the 'system' and the 'builtin' clock sources should be put
into a set of dedicated scenarios.  That's partially so even now: the
scenarios which require verification of the built-in NTP client
functionality are separated into a dedicated test set.  However, it's
necessary to add dedicated test scenarios for the 'system' clock source
and add more coverage for the 'built-in' clock source.

That's exactly what follow-up changelists will do, i.e. add dedicated
test scenarios to cover the specific functionality attributed to the use
of the following clock sources:
  * system  : local machine's wallclock clock synchronized with NTP
  * builtin : independently tracked true time by the built-in NTP client

Change-Id: If7edbf884afaa19121aa92a4ce93c8a7eeb2d937
Reviewed-on: http://gerrit.cloudera.org:8080/14379
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>