KUDU-2069 p8: support listing tserver state through CLI
This patch adds support to the tserver list tool to show the tserver
states.

$ kudu tserver list <master> --columns=uuid,state

I opted to have the state be an optional return field because there are
cases where a tablet that is unregistered will have a state associated
with it (e.g. if the master is restarted), in which case it may be
surprising to see info about them, unless explicitly requested.

Change-Id: I5917927f3bc4bb6bebd1cfb63555ec20d94d9091
Reviewed-on: http://gerrit.cloudera.org:8080/14369
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins