KUDU-2069 p9: add timestamps to tserver states
This patch include timestamps to each tserver states so an operator can
know when a given tserver state was enacted. This patch plumbs this into
the Master's web UI's "Tablet Servers" page.

Here is an example of the Master's web UI with tservers of various
statuses:
https://imgur.com/kny5vHu

When the states are removed, the new section in the web UI does not show
up:
https://imgur.com/itWD0DB

Change-Id: I1b437ad38ad6abcaae894a782ab6d52857c97741
Reviewed-on: http://gerrit.cloudera.org:8080/14370
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins