master: fix the value of table metric 'live_row_count'
When the tablet doesn't support live row counting, the
'live_row_count' metric of the tablet is '-1'. But when
master aggregates all tablets' metric of a table, it sums
up all '-1's, and the result value is confusing to users.
It's better to set it '-1' to be consistent with tablet metric.

Change-Id: I3f837466a0420dafef4f9426929b7b1622702ffc
Reviewed-on: http://gerrit.cloudera.org:8080/14446
Tested-by: Kudu Jenkins
Reviewed-by: Yingchun Lai <405403881@qq.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>