kudu-tool-test: deflake TestFsAddRemoveDataDirEndToEnd
This test was made somewhat flaky by KUDU-2901. The reason is subtle: if any
unrelated IO (such as from a concurrently running test) interleaves with the
creation of the second table, it's possible for the new data directory to be
measured with less available space than the existing ones. The existing LBM
containers have all been preallocated by this point so the test never
remeasures available space. Thus, when the second table flushes, the
KUDU-2901 heuristic ensures that all new data blocks land on the existing
data directories rather than the new one.

The fix is simple: a feature flag for this feature (which we should have had
anyway), which we disable for the test.

Change-Id: Ifb92759318c12d791d26eb8fb6b77914c284f4cb
Reviewed-on: http://gerrit.cloudera.org:8080/14462
Reviewed-by: Andrew Wong <awong@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>
(cherry picked from commit 7e8f37f8ddf46b965b0af4eba3f31c02b11b8fba)
Reviewed-on: http://gerrit.cloudera.org:8080/14469
Tested-by: Kudu Jenkins
Reviewed-by: Grant Henke <granthenke@apache.org>