master: GetTableStatistics should use signed ints
The statistics are reported to the master as signed ints, so the master
should likewise advertise them as such. Otherwise the live_row_count value
of -1 (when a tablet doesn't support it) shows up as:

  TABLE test-workload
  on disk size: 0
  live row count: 18446744073709551615

The public APIs were already treating them as signed; they were only only
unsigned on the wire.

Change-Id: I398cb85888bdc9463264692911788de2fef67dfc
Reviewed-on: http://gerrit.cloudera.org:8080/14463
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>