[spark] Separate out DefaultSourceTests
We have seen flaky test failures due to timeouts of DefaultSourceTest.
This is primarily due to the sheer number of tests in that class.

This patch break out the SQL based tests, ones using `sqlContext.sql(…)`
into their own class. There is no change in test methods or coverage.

The result is 22 DefaultSourceTests and 22 SparkSQLTests.

Change-Id: I54aa0327ffb5254c03fcfe8a0a08dba230360a40
Reviewed-on: http://gerrit.cloudera.org:8080/14491
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Hao Hao <hao.hao@cloudera.com>
Tested-by: Kudu Jenkins