[nvm_cache] remove allocation retry logic
This patch addresses the issue with allocation retry logic
in NvmLRUCache::AllocateAndRetry(), removing it altogether.

The idea of removing some entries from the cache and trying again
in case of allocation failure is:
  * not aligned with DRAM-based cache behavior
  * not robust enough

First, from the conceptual point, DRAM block cache doesn't retry
in such cases (i.e. in the sense that new[] may raise an exception and
crash Kudu if it runs out of memory).  Second, there is no guarantee
there would be some entries in the cache not referenced from anywhere
else.  Third, since the block cache performs clean-up of the cache
on its own by running a periodic grooming task, the value of such
de-allocation attempt is greatly diminished.

This is a follow-up to 0897b6b0e47a64c0cf30c956c020371258f2bacd.

Change-Id: I5671b2413160ca10b7e41c38d7e6164efdccfb8b
Reviewed-on: http://gerrit.cloudera.org:8080/14498
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins