consensus_peers: make RpcPeerProxy::StartElection async and other cleanup
It was the only such proxy call that was synchronous. Apart from violating
POLA, the synchronicity meant that the graceful leadership transfer code
path occupied a Raft threadpool slot for longer in order to receive the RPC
response. It's not a hot path so this is by no means a perf improvement, but
it just seemed unnecessary given that it was only used to log a warning in
the rare event that the remote couldn't start an election.

Changing this meant adding some extra lifecycle code to Peer::StartElection,
but I think it's still a net improvement in clarity.

Along the way I also did some other C++11 cleanup.

Change-Id: I6a6df610f1c07adae5a85534d8c6dec324801042
Reviewed-on: http://gerrit.cloudera.org:8080/14512
Tested-by: Kudu Jenkins
Reviewed-by: Andrew Wong <awong@cloudera.com>