KUDU-2979: Add wrapper function of krb5_parse_name to be used in Impala
From IMPALA-7504, we want to use krb5_parse_name() to parse the principal
instead of using custom code.

When kerberos is initialized in Impala's copy of Kudu code, it stores a
global context which is used when accessing the Krb5 library. To use
this global context, the code that parses the principal name is moved
into Kudu code. This new code is then called from the existing
ParseKerberosPrincipal method.

Test added in mini_kdc-test.TestBasicOperation.

Change-Id: Ifddafa7aae25d66ed7d9fa0306f17501a191cdac
Reviewed-on: http://gerrit.cloudera.org:8080/14520
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>