KUDU-2962 Fix kudu::itest::FindTabletFollowers() test utility function
Fixes the FindTabletFollowers() test utility function to query
each tablet server and form the list of followers instead
of simply finding the leader replica and returning supplied
tablet servers with the leader replica removed.

Earlier implementation is incorrect when one or more of the
supplied tablet servers does not host the specified tablet_id.

Change-Id: I94146873b09cb95c54175d16305b713365cfeaa8
Reviewed-on: http://gerrit.cloudera.org:8080/14533
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins