Adding a StartConsensusOnlyRoundHandler to mimic StartFollowerTransaction
Summary: This is a call to a derived round handler in sync mode
which enables it to reject the update (say if the log is not ready)
and also to do additional bookkeeping (e.g. for starting to buffer
post rotate events )

Test Plan: Tested it inside mysql mtr for Live Master promotion.
Without this, we were not able to stop events after rotate events
to go to the same file.

Reviewers: iRitwik, bhatvinay

Subscribers:

Tasks:

Tags: