schema.h: fix compilation with old gcc
This file fails to compile with the system gcc found on CentOS 6.6
(version 4.4.7) with the following error:

  kudu/client/schema.h:115: error: a class-key must be used when declaring a friend
  kudu/client/schema.h:115: error: friend declaration does not name a class or function
  kudu/client/schema.h:116: error: a class-key must be used when declaring a friend
  kudu/client/schema.h:116: error: friend declaration does not name a class or function
  kudu/client/schema.h:117: error: a class-key must be used when declaring a friend
  kudu/client/schema.h:117: error: friend declaration does not name a class or function

It's a regression from commit 4343fb746. What's perplexing is that while
it's reliably showing up in the dist-test dashboard as a
client_examples-test.sh failure, it's not reliably reproducible. When I'm
logged into a CentOS 6.6 machine I can repro the failed g++ attempt, but if
I run the failed test itself, it passes.

Change-Id: I5916501fc52910bcc0e843726a1d684be6d56417
Reviewed-on: http://gerrit.cloudera.org:8080/14557
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins