thirdparty: add gumbo and gumbo-query
In a follow-on patch I built a simple web crawler for testing the web UI.
It's possible to parse HTML for links via string::find, it made for some
really ugly and brittle code. So I went searching for a proper HTML parser.

I settled on Gumbo (known as gumbo-parser on github), a Google C library for
parsing HTML 5. Although it is quite old and hasn't been updated in some
time, it has been used on Google's web cache and has passed Google's
internal security review. Plus it has an intuitive API.

To simplify further, I incorporated the gumbo-query C++ library, which adds
a CSS selector [2] API for Gumbo. This drastically simplifies an operation
like finding all the links in a page. Sample code:

  string page;
  gq::CDocument doc;
  doc.parse(page);
  gq::CSelection sel = doc.find("a");
  for (int i = 0; i < sel.nodeNum(); i++) {
    string link = sel.nodeAt(i).attribute("href");
    <do stuff with link>
  }

Like Gumbo, gumbo-query is old and unmaintained. I had to patch it to
move all of its functionality into a namespace.

1. https://opensource.googleblog.com/2013/08/gumbo-c-library-for-parsing-html.html
2. https://www.w3schools.com/cssref/css_selectors.asp

Change-Id: I851d5a78e30c1c79a2de253f59c7377a61aa3ed7
Reviewed-on: http://gerrit.cloudera.org:8080/14572
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>