[util] Fix comment of faststring
1. Under any case, capacity() is greater than or equal to
   the data length. If the current length is shorter than
   kInitialCapacity, after call shrink_to_fit(),
   capacity() will go down to kInitialCapacity.
2. The caller must be responsible for releasing the pointer
   release() returns.

Now In `GrowByAtLeast()`, to_reserve (i.e. `len_ + count`)
is usually larger than we really need. This commit also
calculate an exact value when we try to expand the faststring.

To be clearer, this commit rename `GrowByAtLeast()` to
`GrowToAtLeast()`, and make the newcapacity as its argument.

Note: If the capacity needs to grow, it still expand by
at least 50% each time.

Change-Id: Ie4bd4f31216fdf5cfbf70526f1358824fbda225a
Reviewed-on: http://gerrit.cloudera.org:8080/14604
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Adar Dembo <adar@cloudera.com>