KUDU-2989. Work around SASL bug when FQDN is >=64 characters
This adds a workaround for an upstream SASL bug which is triggered when
the FQDN has more than 64 characters. In this case, SASL would truncate
the FQDN and not be able to find the relevant keytab.

The workaround simply uses our own code to determine the FQDN.

Change-Id: I4898814f2f7ab87151798336414dde7078d28a4a
Reviewed-on: http://gerrit.cloudera.org:8080/14609
Reviewed-by: Anurag Mantripragada <anurag@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
(cherry picked from commit 111b13775193820b3e3551368fe00a8f00387007)
Reviewed-on: http://gerrit.cloudera.org:8080/14687
Reviewed-by: Grant Henke <granthenke@apache.org>
Tested-by: Alexey Serbin <aserbin@cloudera.com>