KUDU-2929: don't do nothing when under memory pressure
When we're under memory pressure, we may exit early from considering
ops if there are none that we know will free memory. This might be
because we don't want to spend more memory performing an op when we are
already under memory pressure if we don't know that op is anchoring
memory (obvious in the case of MRS or DMS flushes, less so in the case
of compactions).

While nice in theory, I don't think this behavior was desirable. Yes, we
want to prioritize ops that definitely free memory over those that might
not, but doing _something_ seems better than nothing.

So rather than returning early if there aren't ops that free memory,
let's continue walking through valid ops (in order of what frees the
most WALs, then what frees the most on-disk data, then what has the best
perf improvement).

Change-Id: I030f9ef379af501fe7bd2f42906ec2f9ea16dbde
Reviewed-on: http://gerrit.cloudera.org:8080/14650
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Andrew Wong <awong@cloudera.com>