[clock] run TestNtpDiagnostics against chrony as well
This patch updates the HybridClockTest.TestNtpDiagnostics scenario to
successfully run at machines where 'chronyd' is used as the system
NTP daemon (i.e. rely on the utility suite from the 'chrony' package).
Prior to this patch, the scenario was able to run only against machines
with 'ntpd' as the system daemon (i.e. relied on the utility suite from
the 'ntp' package).

Change-Id: If9c14e939c2722304ce6c5a9f7fc812af183b353
Reviewed-on: http://gerrit.cloudera.org:8080/14751
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins