KUDU-3008 Spread replicas evenly with 2 locations and odd replica factors
In case of distributing 3 replicas among 2 locations, 1 + 2 is better
than 3 + 0 because if all servers in the first location become unavailable,
in the former case the tablet is still available, while in the latter it's
not. Also, 1 + 2 is better than 0 + 3 because in case of catastrophic
non-recoverable failure of the second location, no replica survives and
all data is lost in the latter case, while in the former case there will
be 1 replica and it may be used for manual data recovery.

Change-Id: If96361f6be686679dd72f08e28d964eae25875d0
Reviewed-on: http://gerrit.cloudera.org:8080/14783
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>