KUDU-3006: deflake RebalanceIgnoredTserversTest.Basic
It takes some time for the newly created replicas to bootstrap
and get their status reported as 'healthy' in the catalog manager,
if running the rebalancing prior to that, some unhealthy replicas
would not be moved and the test would fail. So we should wait for
the cluster become healthy and then run the rebalancing.

Change-Id: Icf4b248a5fc91dac0dd6303da08f3b304878bfa2
Reviewed-on: http://gerrit.cloudera.org:8080/14784
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>