KUDU-2992: Avoid sending duplicated requests in catalog_manager
When the master sent a large number of rpc requests to a tserver,
and the tserver didn't response immediately, catalog manager may
continue to send duplicate requests in ProcessTabletReport(), that
would cause rpc timeouts and high load on both masters and tservers.

This patch add a tablet_id index to pending_tasks, catalog manager
would not send duplicate requests to a tablet with the same alter type
when handling a tablet report from a tablet server. If the existing
RPC task failed, catalog manager would create it the next time the
tablet server heartbeats.

This patch also fix HandleResponse() in AsyncDeleteReplica to avoid
unnecessary retries.

Change-Id: If090723821bd78e14a3c54a35cb5e471320002e9
Reviewed-on: http://gerrit.cloudera.org:8080/14849
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>