webserver-crawl-itest: reduce test load
The flaky test dashboard shows a few tests failing with:

  /.../src/kudu/integration-tests/webserver-crawl-itest.cc:299: Failure
  Failed
  Bad status: Network error: curl error: SSL connect error

This appears to be happening in ASAN and TSAN tests. I'm not sure why
anything is timing out as there's no curl timeout configured. In any case,
I've added a SCOPED_TRACE so we can see whether it's always the same URL
that's failing, and I've also removed the read threads from the TestWorkload
as they don't serve any purpose in the test (apart from generating load).

Change-Id: I0b59e9bcf418cde51cb531d7c0d5b38d5f13cd47
Reviewed-on: http://gerrit.cloudera.org:8080/14815
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: Kudu Jenkins